requirements.txt
absl-py==1.2.0
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
anyio==3.6.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.0
beautifulsoup4==4.12.2
black==22.3.0
cachetools==5.2.0
certifi==2022.9.24
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
cloudpickle==2.2.0
Cython==0.29.32
dataclasses==0.6
docker-pycreds==0.4.0
fairscale==0.4.11
fastapi==0.85.0
fasttext==0.9.2
ffmpy==0.3.0
filelock==3.8.0
frozenlist==1.3.1
fsspec==2022.8.2
ftfy==6.1.1
future==0.18.2
fvcore==0.1.5.post20220512
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.12.0
google-auth-oauthlib==0.4.6
gradio==3.4.0
grpcio==1.49.1
h11==0.12.0
h5py==3.9.0
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.10.0
hydra-core==1.2.0
imagecorruptions==1.1.2
imageio==2.29.0
iopath==0.1.9
Jinja2==3.1.2
joblib==1.2.0
lazy_loader==0.2
linkify-it-py==1.0.3
llvmlite==0.40.0
lvis==0.5.3
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.7.2
mdit-py-plugins==0.3.1
mdurl==0.1.2
mmcv-full==1.6.0
mss==6.1.0
multidict==6.0.2
munkres==1.1.4
mypy-extensions==0.4.3
networkx==3.1
nltk==3.7
numba==0.57.0
numpy==1.23.3
oauthlib==3.2.1
omegaconf==2.2.3
opencv-python==4.6.0.66
orjson==3.8.0
pandas==1.5.0
paramiko==2.11.0
pathspec==0.10.1
pathtools==0.1.2
Pillow==9.4.0
platformdirs==2.5.2
ply==3.11
portalocker==2.5.1
protobuf==3.19.6
psutil==5.9.5
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.12.0
pycocotools==2.0.5
pycryptodome==3.15.0
pydantic==1.10.2
pydot==1.4.2
pydub==0.25.1
PyNaCl==1.5.0
PyQt5-sip==12.11.0
python-multipart==0.0.5
pytz==2022.4
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.9.13
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.1.2
scipy==1.9.1
sentry-sdk==1.24.0
setproctitle==1.3.2
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.20.4
tabulate==0.8.10
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.0.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.5.4
tomli==2.0.1
torch==1.11.0
torch-kmeans==0.2.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.1
uc-micro-py==1.0.1
uvicorn==0.18.3
wandb==0.15.3
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.2.2
yacs==0.1.8
yapf==0.32.0
yarl==1.8.1
